Created chat wrapper
