+++ src/feature/chat/components/chat-id/ChatId.tsx
@@ -1,21 +1,14 @@
-import { Div } from '@/components/atoms'
 import React from 'react'
-import { MainContainer, SecondaryBackground } from '../../../../layouts/wrapper'
 import ChatPreviewList from '../chat/chat-preview-list'
+import ChatWrapper from '../chat/chat-wrapper'
 import Message from '../chat/message'
-import useAppQuery from '../../../../hooks/useAppQuery'
 
 const ChatId: React.FC = () => {
-    const query = useAppQuery('chat-wrapper')
     return (
-        <MainContainer>
-            <SecondaryBackground>
-                <Div query={query}>
-                    <ChatPreviewList />
-                    <Message />
-                </Div>
-            </SecondaryBackground>
-        </MainContainer>
+        <ChatWrapper>
+            <ChatPreviewList />
+            <Message />
+        </ChatWrapper>
     )
 }
