+++ src/feature/chat/components/chat/chat-wrapper/index.ts
@@ -1,0 +1,1 @@
+export {default} from './ChatWrapper'
