+++ src/feature/chat/components/chat/message/Message.tsx
@@ -1,16 +1,20 @@
 import React from 'react'
-import { P } from '@/components/atoms'
+import { Div, P } from '@/components/atoms'
 import { SecondaryContainer } from '../../../../../layouts/wrapper'
 import MessageCreator from './message-creator'
 import MessageList from './message-list'
+import useAppStyles from 'src/hooks/useAppStyles'
 
 const Message: React.FC = () => {
+    const className = useAppStyles('children/height-auto')
     return (
-        <SecondaryContainer>
-            <P>Name of chat / user</P>
-            <MessageList />
-            <MessageCreator />
-        </SecondaryContainer>
+        <Div className={className}>
+            <SecondaryContainer>
+                <P>Name of chat / user</P>
+                <MessageList />
+                <MessageCreator />
+            </SecondaryContainer>
+        </Div>
     )
 }
